requirements.txt
gstreamer-python @ git+https://github.com/jackersson/gstreamer-python.git